lambda/requirements.txt
requests==2.32.3
openai==1.61.0
pdfplumber==0.11.5
boto3==1.36.11
markdown==3.7
python-markdown-math==0.8
importlib-metadata==8.6.1
pygments==2.19.1
pymdown-extensions==10.14.3
markdown-it-py[linkify,plugins]